requirements.txt
asgiref==3.3.1
autopep8==1.5.6
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.1.2
pycodestyle==2.7.0
PyJWT==2.0.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
